Add admin dashboard
As the very first step building the admin area, I have added the dashboard.
It doesn't show much. Only the main menu and a welcome message. This can be
extended in the future, but it's fine in the way it is right now.

This version hasn't any user permission check included, meaning that everyone
can see the admin area, no matter whether the user has signed in or not.