Add credits page
Added credits page to list all the libraries and frameworks
I have used to build this blog system.

Also changed the location for all page templates. Before the login
template was inside its own login folder. But that made not much
sense in terms of folder structure. So, I have changed it to a
central location for all page templates. This makes the structure
much easier.