Fix login issue in case the user is already signed in
In case you're already signed in and visit the login page,
you have to sign in again, but actually u're still authenticated.
I have added a check that will lead to a redirection to the admin
dashboard in case you're still signed in otherwise you keep on the
login page.